pyproject.toml
[project]
name = "service-concierge-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.26.0",
]
